CMakeLists.txt done

+++ src/basic/layer.h
@@ -10,8 +10,8 @@ class Layer {
 public:
     int id; 
     bool direction; // 0: horizontal, 1: vertical
-    int minLength;
-    vector<vector<int>> capacity; 
+    double minLength;
+    vector<vector<double>> capacity; 
     
 };
